Added client validation invite registration form.
