+++ client/src/utils/auth.js
@@ -5,6 +5,7 @@ import api from "./api";
 import { useStoreContext, getStoreAction as gsa } from "store";
 import { LOGIN_USER, LOGOUT_USER } from "store/actions";
 import { useHistory } from "react-router-dom";
+import { createValidator } from "components/Form";
 
 const setAuthToken = token => {
 
@@ -171,3 +172,35 @@ export const useLogout = () => {
     }
     
 }
+
+export const validateRegistrationData = createValidator({
+    validators: {
+        name: ({ name }) => Boolean(name) || "Your name is required",
+        email: ({ email }) => Boolean(email) || "Email is required",
+        password: ({ password, password2 }) => {
+
+            const errors = [];
+
+            if( password ) {
+
+                // TODO stronger password validation for both client and server.
+                if( password.length < 6 ) errors.push(["password", "Password must be at least 6 characters"] );
+
+            } else {
+                errors.push( ["password", "Password is required"] );
+            }
+
+            if( !password2 ) errors.push( ["password2", "Confirm password is required"] );
+
+            if( !errors.length ) {
+                
+                if( password !== password2 ) errors.push( ["password", "Passwords must match"], ["password2", "Passwords must match"] );
+
+            }
+
+            if( errors.length ) return Object.fromEntries( errors );
+
+        },
+        roomname: ({ roomname }) => Boolean(roomname) || "Room name is required",
+    }
+});
